// Project-RUPP/nc-visal-shop/src/assets/components/interfaces/ProductType.tsx
export interface Product {
    id: string;
    img: string;
    gallery: string[];
    title: string;
    stars: string;
    brand: string;
    reviews: string;
    shortDes: string;
    des: string;
    prevPrice: number;
    salePrice: number;
    qty: number;
    company: string;
    category: string;
    model: string;
    size: string[];
  }
  